Refactored server start code into prod, dev, test modes
This change makes the lifecycle for starting up applications much
clearer.

* No longer need separate ServerStart implementations for Netty and
  Akka HTTP because ServerProvider configuration is always loaded from
  configuration files. Instead, separate out code according to the mode
  that the server runs in, because behavior can vary between modes. Now
  we have a ProdServerStart, DevServerStart and a DocServerStart.
* For each mode, move the ApplicationProvider code into same file as
  the new server startup code. Move code for starting up the application
  out of the ApplicationProvider constructors and into the server
  start code. ApplicationProviders still implement the 'get' method
  for getting the current Application.
* Remove TestApplication and StaticApplication, because they do the
  same thing. Instead provide helpers for 'static' Applications that
  don't need reloading.
