+++ framework/src/play-akka-http-server/src/test/scala/play/core/server/akkahttp/AkkaHttpServerSpec.scala
@@ -25,7 +25,7 @@ object AkkaHttpServerSpec extends PlaySpecification with WsTestClient {
         routes: PartialFunction[(String, String), Handler])(
           check: WSResponse => T)(
             implicit awaitTimeout: Timeout): T = {
-    running(TestServer(testServerPort, FakeApplication(withRoutes = routes), serverProvider = Some(AkkaHttpServer.defaultServerProvider))) {
+    running(TestServer(testServerPort, FakeApplication(withRoutes = routes), serverProvider = Some(AkkaHttpServer.provider))) {
       val plainRequest = wsUrl(path)(testServerPort)
       val responseFuture = exec(plainRequest)
       val response = await(responseFuture)(awaitTimeout)
@@ -149,7 +149,7 @@ object AkkaHttpServerSpec extends PlaySpecification with WsTestClient {
 
     "support WithServer form" in new WithServer(
       app = FakeApplication(withRoutes = httpServerTagRoutes),
-      serverProvider = Some(AkkaHttpServer.defaultServerProvider)) {
+      serverProvider = Some(AkkaHttpServer.provider)) {
       val response = await(wsUrl("/httpServerTag").get())
       response.status must equalTo(OK)
       response.body must_== "Some(akka-http)"
@@ -162,7 +162,7 @@ object AkkaHttpServerSpec extends PlaySpecification with WsTestClient {
           val app = FakeApplication(withRoutes = {
             case ("GET", "/") => Action(Ok(resultString))
           })
-          val server = TestServer(testServerPort, app, serverProvider = Some(AkkaHttpServer.defaultServerProvider))
+          val server = TestServer(testServerPort, app, serverProvider = Some(AkkaHttpServer.provider))
           server.start()
           try {
             val response = await(wsUrl("/")(testServerPort).get())
