+++ framework/src/play-server/src/main/scala/play/core/server/ssl/DefaultSSLEngineProvider.scala
@@ -55,7 +55,7 @@ class DefaultSSLEngineProvider(serverConfig: ServerConfig, appProvider: Applicat
     } else {
       // Load a generated key store
       logger.warn("Using generated key with self signed certificate for HTTPS. This should not be used in production.")
-      FakeKeyStore.keyManagerFactory(applicationProvider.path)
+      FakeKeyStore.keyManagerFactory(serverConfig.rootDir)
     }
 
     // Load the configured trust manager
