+++ framework/src/play-server/src/test/resources/application.conf
@@ -1,0 +1,1 @@
+# Needed so play-server tests run
