+++ framework/src/sbt-plugin/src/sbt-test/play-sbt-plugin/distribution/build.sbt
@@ -18,7 +18,7 @@ checkStartScript := {
   }
   val contents = IO.read(startScript)
   val lines = IO.readLines(startScript)
-  if (!contents.contains( """app_mainclass="play.core.server.NettyServer"""")) {
+  if (!contents.contains( """app_mainclass="play.core.server.ProdServerStart"""")) {
     startScriptError(contents, "Cannot find the declaration of the main class in the script")
   }
   val appClasspath = lines.find(_ startsWith "declare -r app_classpath")
